requirements.txt
flask>=3.0.0
line-bot-sdk>=3.5.0
openai>=1.30.0
settrade-v2==2.2.1
pandas>=2.0.0
requests>=2.31.0
gunicorn>=21.2.0
matplotlib>=3.8.0
numpy>=1.26.0
